Vehicle Search Form
Allows user to search for a vehicle in the inventory by querying the inventory database. If vehicles are found matching that criteria they will be loaded into list box and then the user can select a vehicle and click view vehicle, which will display a form with all the vehicle specifications.